% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{callback_save_model_state_dict}
\alias{callback_save_model_state_dict}
\title{callback_save_model_state_dict}
\usage{
callback_save_model_state_dict(save_dir = ".", prefix = "model_epoch")
}
\arguments{
\item{save_dir}{path and directory to save state dictionary files}

\item{prefix}{prefix for file name. Default is "model_epoch". The suffix
is the epoch number.}
}
\value{
No R object is returned. The state dictionary is saved to disk as a .pt file.
}
\description{
Save the model state dictionary to disk at the end of each epoch as a .pt file.
For use within luz training loop.
}
