% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{countParams}
\alias{countParams}
\title{countParams}
\usage{
countParams(model, test_data)
}
\arguments{
\item{model}{instantiated model object as nn_module subclass as opposed to luz fitted object.}

\item{test_data}{Example tensor of the correct shape for the model being explored.}
}
\value{
Counts of model parameters.
}
\description{
Count the number of trainable parameters in a model.
}
\details{
Count the number of trainable parameters in an instantiated model subclassed from nn_module().
}
