% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBatch.R
\name{describeBatch}
\alias{describeBatch}
\title{describeBatch}
\usage{
describeBatch(dataLoader, usedDS = FALSE)
}
\arguments{
\item{dataLoader}{Instantiated instance of a DataLoader created using torch::dataloader().}

\item{usedDS}{TRUE or FALSE. If deep supervision is implemented, this must be set to TRUE.
Default is FALSE. Whether or not deep supervision was used, only metrics for the masks
at the original spatial resolution are returned.}
}
\value{
List object summarizing a batch of image chips and masks.
}
\description{
Generate summary information for a batch of image chips and masks.
}
\details{
The goal of this function is to provide a check of a batch of image chips and associated
masks generated by a DataLoader instance using defineSegDataSet(). Summary information includes the batch size
(batchSize); image chip data type (imageDataType); mask data type (maskDataType); the shape of the batch of images
or predictor variables as mini-batch size, number of channels, height pixel count, and width pixel count (imageShape); the mask shape
(maskShape); image band means (bndMns); image band standard deviations (bndSDs);
count of pixels in each class in the batch (mskCnts); and minimum (minIndex) and maximum (maxIndex) class indices present in the batch.
}
