% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luzMetrics.R
\name{luz_metric_accuracyDS}
\alias{luz_metric_accuracyDS}
\title{luz_metric_accuracyDS}
\usage{
luz_metric_accuracyDS()
}
\value{
Calculated metric return as a base-R vector as opposed to tensor.
}
\description{
luz_metric function to calculate the overall accuracy for a multiclass classification
}
\details{
Calculates overall accuracy for a multiclass classification. This is an augmentation of the
original luz_metric_accuracy() function from luz that has been modified to support deep supervision.
It should only be used if deep supervision is implemented. If deep supervision is not implemented,
use the original implementation from luz.
}
