% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luzMetrics.R
\name{luz_metric_binary_accuracy_with_logitsDS}
\alias{luz_metric_binary_accuracy_with_logitsDS}
\title{luz_metric_binary_accuracy_with_logitsDS}
\usage{
luz_metric_binary_accuracy_with_logitsDS(threshold = 0.5)
}
\value{
Calculated metric return as a base-R vector as opposed to tensor.
}
\description{
luz_metric function to calculate the overall accuracy for a binary classification from predicted logits
}
\details{
Calculates overall accuracy for a binary classification. This is an augmentation of the
original luz_metric_binary_accuracy_with_logits() function from luz that has been modified
to support deep supervision. It should only be used if deep supervision is implemented.
If deep supervision is not implemented, use the original implementation from luz.
}
