% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luzMetrics.R
\name{luz_metric_recall}
\alias{luz_metric_recall}
\title{luz_metric_recall}
\usage{
luz_metric_recall(mode = "multiclass", average = "micro", smooth = 1)
}
\arguments{
\item{mode}{Either "binary" or "multiclass". If "binary", only the logit for
positive class prediction should be provided. If both the positive and negative
or background class logits are provided for a binary classification, use
the "multiclass" mode.}

\item{average}{Either "micro" or "macro". Whether to use micro- or macro-averaging
for multiclass metric calculation. Ignored when mode is "binary". Default is
"micro"}

\item{preds}{Tensor of class predicted probabilities with shape
(Batch, Class Logits, Height, Width) for a multiclass classification. For a
binary classification, you should provide logits for just the positive class as
(Batch, Positive Class Logit, Height, Width) or (Batch, Height, Width).}

\item{target}{Tensor of target class labels with shape
(Batch, Class Indices, Height, Width) for a multiclass classification. For a
binary classification, you should provide targets as
(Batch, Positive Class Index, Height, Width) or (Batch, Height, Width). For
binary classification, the class index must be 1 for the positive class and 0 for the
background case.}
}
\value{
Calculated metric return as a base-R vector as opposed to a tensor
}
\description{
luz_metric function to calculate recall
}
\details{
Calculates recall based on luz_metric() for use within training and validation
loops.
}
