% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChipsDF.R
\name{makeChipsDF}
\alias{makeChipsDF}
\title{makeChipsDF}
\usage{
makeChipsDF(
  folder,
  outCSV,
  extension = ".tif",
  mode = "All",
  shuffle = FALSE,
  saveCSV = FALSE
)
}
\arguments{
\item{folder}{Full path or path relative to the working directory to the
folder containing the image chips and associated masks. You must include the
final forward slash in the path (e.g., "C:/data/chips/").}

\item{outCSV}{File name and full path or path relative to the working directory
for the resulting CSV file with a ".csv" extension.}

\item{extension}{The extension of the image and mask raster data (e.g., ".tif",
".png", ".jpeg", or ".img"). The default is ".tif" since this is the file
format used by the utilities in this package. This option is provided if chips
are generated using another method.}

\item{mode}{Either "All", "Positive", or "Divided". This should match the setting
used in the makeChips() function. If the makeChipsMultiClass() function was used,
this should be set to "All" or left as the default. The default is "All".}

\item{shuffle}{TRUE or FALSE. Whether or not to shuffle the rows in the table.
Rows can be shuffled to potentially reduced autocorrelation in the data. The
default is FALSE.}

\item{saveCSV}{TRUE or FALSE. Whether or not to save the CSV file or just
return the data frame. If this is set to FALSE then the outCSV parameter is
ignored and no CSV file is generated. The default is FALSE.}
}
\value{
Data frame with three columns (chpN, chpPth, and mskPth) and, optionally,
a CSV file written to disk. If mode = "Divided", a division column is added to
differentiate "positive" and "background" samples.
}
\description{
Create data frame and CSV file listing image chips and associated masks
}
\details{
This function creates a data frame and, optionally, a CSV file that lists all
of the image chips and associated masks in a directory. Three columns are
produced. The chpN column provides the name of the chip, the chpPth column
provides the path to the chip, and the chpMsk column provides the path to the
associated mask. All paths are relative to the input folder as opposed to
the full file path so that the results can still be used if the data are copied
to a new location on disk or to a new computer.
}
