% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeBatch.R
\name{describeBatch}
\alias{describeBatch}
\title{describeBatch}
\usage{
describeBatch(dataLoader, usedDS = FALSE)
}
\arguments{
\item{dataLoader}{Instantiated instance of a DataLoader created using torch::dataloader().}
}
\value{
List object summarizing a batch of image chips and masks.
}
\description{
Generate summary information for a batch of image chips and masks.
}
\details{
The goal of this function is to provide a check of a batch of image chips and associated
masks generated by a DataLoader instance using defineSegDataSet(). Summary information includes the batch size
(batchSize); image chip data type (imageDataType); mask data type (maskDataType); the shape of each
image chip as number of channels, height pixel count, and width pixel count (imageShape); the mask shape
(maskShape); image band statistics (imageStats) including mean (mean), median (median), minimum (min),
maximum (max), and standard deviation (sd); mask class index statistics (mskStats) including minimum (min)
and maximum (max) class indices; and count of pixels in each class in the batch (mskCnts).
}
