% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luzMetrics.R
\name{luz_metric_f1score}
\alias{luz_metric_f1score}
\title{luz_metric_f1score}
\usage{
luz_metric_f1score(
  nCls = 1,
  smooth = 1,
  mode = "multiclass",
  biThresh = 0.5,
  zeroStart = TRUE,
  usedDS = FALSE
)
}
\arguments{
\item{smooth}{a smoothing factor to avoid divide by zero errors. Default is 1.}

\item{mode}{Either "binary" or "multiclass". If "binary", only the logit for
positive class prediction should be provided. If both the positive and negative
or background class probability is provided for a binary classification, use
the "multiclass" mode.}

\item{zeroStart}{TRUE or FALSE. If class indices start at 0 as opposed to 1, this should be set to
TRUE. This is required  to implement one-hot encoding since R starts indexing at 1. Default is TRUE.}

\item{preds}{Tensor of class predicted probabilities with shape
(Mini-Batch, Class Logits, Width, Height) and a 32-bit float data type.}

\item{target}{Tensor of target class labels with shape
(Mini-Batch, Class Indices, Width, Height) and a long integer data type. For
binary classification, the class index must be 1 for the positive class and
0 for the background case.}

\item{nCLs}{number of classes being differentiated.}
}
\value{
Calculated metric returned as a base-R vector as opposed to tensor.
}
\description{
luz_metric function to calculate the F1-score
}
\details{
Calculates F1-score based on luz_metric() for use within training and validation
loops.
}
