% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeChipsMultiClass.R
\name{makeChipsMultiClass}
\alias{makeChipsMultiClass}
\title{Generate image chips from images and associated raster masks for multiclass classification}
\usage{
makeChipsMultiClass(
  image,
  mask,
  n_channels = 3,
  hasZero = TRUE,
  size = 256,
  stride_x = 256,
  stride_y = 256,
  outDir
)
}
\arguments{
\item{image}{Path to input image. Function will generate a SpatRaster object
internally. The image and mask must have the same extent, number of rows and
columns of pixels, cell size, and coordinate reference system.}

\item{mask}{Path to single-band mask. Function will generate a SpatRaster
object internally. The image and mask must have the same extent, number of
rows and columns of pixels, cell size, and coordinate reference system.}

\item{n_channels}{Number of channels in the input image. Default is 3.}

\item{hasZero}{If the class codes begin at 0 as opposed to 1, this shouls be
set to TRUE. If the class codes start at 1 as opposed to 0, this should be set
to FALSE. In the case where the class codes start at 1, each code will be
reduced by 1 so that codes start at 0. For example, codes 1,2,3,4 would be
converted to 0,1,2,3. This is because most deep learning frameworks expect the
class codes to be represented as indices from 0 to n-1 where n is the number
of classes. Users should be aware of this manipulation and its impact of class
code mappings.}

\item{size}{Size of image chips as number of rows and columns of pixels.
Default is 256.}

\item{stride_x}{Stride in the x (columns) direction. Default is 256.}

\item{stride_y}{Stride in the y (rows) direction. Default is 256.}

\item{outDir}{Full or relative path to the current working directory where you
want to write the chips to. Subfolders in this directory will be generated by
the function. You must include the final forward slash in the file path
(e.g., "C:/data/chips/").}
}
\value{
Image and mask files written to disk in TIFF format. No R object is
returned.
}
\description{
This function will generate image and mask chips from an input image and
associated raster mask. The chips will be written into the defined directory.
The number of rows and columns of pixels per chip are equal to the size argument.
If a stride_x and/or stride_y is used that is different from the size argument,
resulting chips will either overlap or have gaps between them. In order to
not have overlap or gaps, the stride_x and stride_y arguments should be the
same as the size argument. Both the image chips and associated masks are
written to TIFF format (".tif"). Input data are not limited to three
band images. This function is specifically for a multiclass classification.
For a binary classification, use the makeChips() function. If an irregular shaped
raster grid is provided, only chips and masks that contain no NA or NoDATA
cells will be produced.
}
\details{
Within the provided directory, image chips will be written to an "images" folder
and masks will be written to a "masks" folder.
}
