% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewBatch.R
\name{viewBatch}
\alias{viewBatch}
\title{viewBatch}
\usage{
viewBatch(
  dataLoader,
  chnDim = TRUE,
  mskLong = TRUE,
  nRows = 3,
  r = 1,
  g = 2,
  b = 3,
  cNames,
  cColors,
  usedDS = FALSE
)
}
\arguments{
\item{dataLoader}{Instantiated instance of a DataLoader created using torch::dataloader().}

\item{chnDim}{TRUE or FALSE. Default is TRUE. If TRUE, assumes the target tensor includes the
channel dimension: (Batch, Channel, Height, Width). If FALSE, assumes the target tensor does not include the
channel dimension: (Channel, Height, Width). The script is written such that it expects the channel dimension.
So, if FALSE, the script will add the channel dimension as needed.}

\item{mskLong}{TRUE or FALSE. Default is TRUE. Data type of target or mask. If the provided target
has a data type other than torch_long, this parameter should be set to FALSE. This will cause the
script to convert the target to the tensor_long data type as required for the calculations.}

\item{nRows}{Number of rows in the image grid. Default is 3.}

\item{r}{Index of channel to assign to red channel. Default is 1.
For gray scale or single-band images, assign the same index to all three bands.}

\item{g}{Index of channel to assign to green channel. Default is 2.
For gray scale or single-band images, assign the same index to all three bands.}

\item{b}{Index of channel to assign to blue channel. Default is 3.
For gray scale or single-band images, assign the same index to all three bands.}

\item{cNames}{Vector of class names. Must be the same length as number of classes.}

\item{cColors}{Vector of color values to use to display the masks. Colors are applied based on the
order of class indices. Length of vector must be the same as the number of classes.}

\item{usedDS}{TRUE or FALSE. Must be set to TRUE when using defineSegDataSetDS(). Default is FALSE,
or it is assumed that deep supervision is not used.}
}
\value{
Image grids of example chips and masks loaded from a batch produced by the DataLoader.
}
\description{
Generate image grid of batch of image chips and associated masks created by a DataLoader.
}
\details{
The goal of this function is to provide a visual check of a batch of image chips and associated masks
generated from a DataLoader.
}
