% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewBatchPreds.R
\name{viewBatchPreds}
\alias{viewBatchPreds}
\title{viewBatchPreds}
\usage{
viewBatchPreds(
  dataLoader = testDL,
  model = model2,
  mode = "multiclass",
  chnDim = FALSE,
  mskLong = TRUE,
  nRows = 4,
  r = 1,
  g = 2,
  b = 3,
  cNames,
  cColors,
  useCUDA = TRUE,
  probs = FALSE,
  usedDS = FALSE
)
}
\arguments{
\item{dataLoader}{Instantiated instance of a DataLoader created using torch::dataloader().}

\item{model}{Fitted model used to predict batch.}

\item{mode}{"multiclass" or "binary". If the prediction returns the postive case logit
for a binary classification problem, use "binary". If 2 or more class logits are returned,
use "multiclass".}

\item{chnDim}{TRUE or FALSE. Default is TRUE. If TRUE, assumes the target tensor includes the
channel dimension: (Batch, Channel, Height, Width). If FALSE, assumes the target tensor does not include the
channel dimension: (Channel, Height, Width). The script is written such that it expects the channel dimension.
So, if FALSE, the script will add the channel dimension as needed.}

\item{mskLong}{TRUE or FALSE. Default is TRUE. Data type of target or mask. If the provided target
has a data type other than torch_long, this parameter should be set to FALSE. This will cause the
script to convert the target to the tensor_long data type as required for the calculations.}

\item{nRows}{Number of rows in the image grid. Default is 3.}

\item{r}{Index of channel to assign to red channel. Default is 1.
For gray scale or single-band images, assign the same index to all three bands.}

\item{g}{Index of channel to assign to green channel. Default is 2.
For gray scale or single-band images, assign the same index to all three bands.}

\item{b}{Index of channel to assign to blue channel. Default is 3.
For gray scale or single-band images, assign the same index to all three bands.}

\item{cNames}{Vector of class names. Must be the same length as number of classes.}

\item{cColors}{Vector of color values to use to display the masks. Colors are applied based on the
order of class indices. Length of vector must be the same as the number of classes.}

\item{useCUDA}{TRUE or FALSE. Default is FALSE. If TRUE, GPU will be used to predict
the data batch. If FALSE, predictions will be made on the CPU.}

\item{probs}{TRUE or FALSE. Default is FALSE. If TRUE, class probabilities will be
shown as opposed to the hard classification. If FALSE, hard classification will be
shown. For a binary problem, the positive class logit is transformed using a sigmoid
function. For multiclass, softmax is used to transform the class logits to probabilities
that sum to 1.}

\item{usedDS}{TRUE or FALSE. Must be set to TRUE when using defineSegDataSetDS(). Default is FALSE,
or it is assumed that deep supervision is not used.}
}
\value{
Image grids of example chips and masks loaded from a batch produced by the DataLoader.
}
\description{
Generate image grid of batch of image chips, masks, and predictions for all samples in a DataLoader batch.
}
\details{
The goal of this function is to provide a visual check of predictions for a batch of data.
}
