% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewChips.R
\name{viewChips}
\alias{viewChips}
\title{viewChips}
\usage{
viewChips(
  chpDF,
  folder,
  nSamps = 16,
  mode = "both",
  justPositive = FALSE,
  cCnt = 4,
  rCnt = 4,
  r = 1,
  g = 2,
  b = 3,
  rescale = FALSE,
  rescaleVal = 1,
  cNames,
  cColors,
  useSeed = FALSE,
  seed = 42
)
}
\arguments{
\item{chpDF}{Data frame of chip paths created with the makeChipsDF() function.}

\item{folder}{Full path or path relative to the working directory to the
folder containing the image chips and associated masks. You must include the
final forward slash in the path (e.g., "C:/data/chips/").}

\item{nSamps}{Number of samples to include in the grid. The default is 16.}

\item{mode}{Either "image", "mask" or "both". If "image", a grid is produced
for the image chips only. If "mask", a grid is produced for just the masks.
If "both", masks are produced for both the image chips and masks. Default is
"both".}

\item{cCnt}{Number of columns in the grid. Row X Column count must sum to the number
of samples being displayed (nSamps). Default is 4.}

\item{rCnt}{Number of rows in the grid. Row X Column count must sum to the number
of samples being displayed (nSamps). Default is 4.}

\item{r}{Band number to map to the red channel. Default is 1.}

\item{g}{Band number to map to the green channel. Default is 2.}

\item{b}{Band number to map to the red channel. Default is 3.}

\item{cNames}{Vector of class names. Class names must be provided.}

\item{cColors}{Vector of colors (named colors, hex codes, or rgb()).
Colors to use to visualize each class matched based on position
in the vector. Colors must be provided.}

\item{useSeed}{TRUE or FALSE. Whether or not to set a random seed to make result
reproducible. If FALSE, seed is ignored. Default is FALSE.}

\item{seed}{Random seed value. Default is 42. This is ignored if useSeed is FALSE.}

\item{justPostitive}{TRUE or FALSE. If makeChips() was executed using the mode "Divided", you can
choose to only show chips that contained some pixels mapped to the positive class.
The default is FALSE. This should be left to the default or set to FALSE if chips
were generated using a method other than "Divided".}
}
\value{
Plot of image chip grid (if mode = "image"); plot of mask chip grid
(if mode ="mask"); plot of image and mask chip grids (if model = "both").
}
\description{
Plot a grid of image and/or mask chips
}
\details{
This function generates a plot of image chips and/or image masks. It serves
as a means to visualize chips generated with the makeChips() or makeChipsMultiClass()
function. It can be used as a check to make sure chips were generated as expected.
}
